src/modules/cars/useCases/createCarSpecification/CreateCarSpecificationUseCase.spec.ts
import { SpecificationsRepositoryInMemory } from '@modules/cars/repositories/in-memory/SpecificationsRepositoryInMemory';
import { AppError } from '@shared/errors/AppError';
import { CreateCarSpecificationUseCase } from './createCarSpecificationUseCase';
import { CarsRepositoryInMemory } from '@modules/cars/repositories/in-memory/CarsRepositoryInMemory';

let carsRepositoryInMemory: CarsRepositoryInMemory;
let specificationsRepositoryInMemory: SpecificationsRepositoryInMemory;

let createCarSpecificationUseCase: CreateCarSpecificationUseCase;

describe('Create Car Specification', () => {
  beforeEach(() => {
    carsRepositoryInMemory = new CarsRepositoryInMemory();
    specificationsRepositoryInMemory = new SpecificationsRepositoryInMemory();

    createCarSpecificationUseCase = new CreateCarSpecificationUseCase(
      carsRepositoryInMemory,
      specificationsRepositoryInMemory
    );
  });

  it('Should be able a add a new specification to now-existent car', async () => {
    expect(async () => {
      await createCarSpecificationUseCase.execute({
        car_id: '1',
        specifications_id: ['2'],
      });
    }).rejects.toBeInstanceOf(AppError);
  });

  it('Should not be able a add a new specification to the car', async () => {
    const car = await carsRepositoryInMemory.create({
      name: 'Name car available',
      description: 'Description car',
      daily_rate: 20,
      license_plate: 'ABC-1234',
      fine_amount: 60,
      brand: 'Brand',
      category_id: 'category',
    });

    const specification = await specificationsRepositoryInMemory.create({
      name: 'test',
      description: 'test',
    });

    const specificationAdded = await createCarSpecificationUseCase.execute({
      car_id: car.id,
      specifications_id: [specification.id],
    });

    expect(specificationAdded).toHaveProperty('specifications');
    expect(specificationAdded.specifications.length).toBe(1);
  });
});
